Fix unit_price data type

+++ db/migrate/20201217193257_create_item.rb
@@ -3,7 +3,7 @@ class CreateItem < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.integer :unit_price
+      t.float :unit_price
       t.bigint :merchant_id
 
       t.timestamps
